pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "flask-wtf>=1.2.2",
    "sqlalchemy>=2.0.40",
    "werkzeug>=3.1.3",
    "wtforms>=3.2.1",
    "markupsafe>=3.0.2",
    "requests>=2.32.3",
    "flask-talisman>=1.1.0",
    "pyotp>=2.9.0",
    "qrcode>=8.0",
    "pillow>=11.1.0",
    "python-telegram-bot>=22.0",
    "python-dotenv>=1.1.0",
    "flask-limiter>=3.12",
    "pytz>=2025.2",
    "sendgrid>=6.11.0",
    "sib-api-v3-sdk>=7.6.0",
]
